# Conflicts:
#	nts.uk/pr/nts.uk.pr.web/src/main/webapp/view/qpp/005/b/start.js
#	nts.uk/pr/nts.uk.pr.web/src/main/webapp/view/qpp/005/b/start.ts
#	nts.uk/pr/nts.uk.pr.web/src/main/webapp/view/qpp/005/b/viewmodel.js
#	nts.uk/pr/nts.uk.pr.web/src/main/webapp/view/qpp/005/b/viewmodel.ts

+++ nts.uk/pr/proto/nts.uk.pr.proto.infra/src/main/java/nts/uk/ctx/pr/proto/infra/entity/paymentdata/QstdtPaymentDetail.java
@@ -69,7 +69,7 @@ public class QstdtPaymentDetail {
 	
 	@Basic(optional = false)
 	@Column(name = "COMMU_ALLOW_MONTH")
-	public int commuteAllowMonth;
+	public BigDecimal commuteAllowMonth;
 	
 	@Basic(optional = false)
 	@Column(name = "COMMU_ALLOW_FRACTION")
